# [AIRFLOW-5931] Use os.fork when appropriate to speed up task execution.