--- a/airflow/utils/dag_processing.py
+++ b/airflow/utils/dag_processing.py
@@ -612,6 +612,9 @@ class DagFileProcessorManager(LoggingMixin):  # pylint: disable=too-many-instanc
         user code.
         """
 
+        # Start a new process group
+        os.setpgid(0, 0)
+
         self.log.info("Processing files using up to %s processes at a time ", self._parallelism)
         self.log.info("Process each file at most once every %s seconds", self._file_process_interval)
         self.log.info(