@dimberman 2019-11-21
I have tested this locally and it seems to work fine. 

@ashb when are situations where CAN_FORK is false besides when doing run_as_user?